Added try block for invalid args.
If a user doesn't supply a number after -p or --port it will print out a
help message and end the program. Also added the -h and --help flags to
also print out a help message and exit.
